requirements.txt
agno
openai
python-dotenv
uvicorn
fastapi
duckduckgo-search
psycopg2-binary
sqlalchemy
asyncpg